+++ src/types/custom_types.rs
@@ -5,15 +5,47 @@ use crate::Identifier;
 use std::prelude::v1::*;
 use std::{any::type_name, collections::BTreeMap};
 
-/// _(internals)_ Information for a custom type.
-/// Exported under the `internals` feature only.
+/// Information for a registered custom type.
 #[derive(Debug, Eq, PartialEq, Clone, Hash, Default)]
 pub struct CustomTypeInfo {
+    /// Rust name of the custom type.
+    pub(crate) type_id: Identifier,
     /// Friendly display name of the custom type.
-    pub display_name: Identifier,
+    pub(crate) display_name: Identifier,
     /// Comments.
     #[cfg(feature = "metadata")]
-    pub comments: Box<[crate::SmartString]>,
+    pub(crate) comments: Box<[crate::SmartString]>,
+}
+
+impl CustomTypeInfo {
+    /// Rust name of the custom type.
+    #[inline(always)]
+    #[must_use]
+    pub fn type_id(&self) -> &str {
+        &self.type_id
+    }
+    /// Friendly display name of the custom type.
+    #[inline(always)]
+    #[must_use]
+    pub fn display_name(&self) -> &str {
+        &self.display_name
+    }
+    /// _(metadata)_ Iterate the doc-comments defined on the custom type.
+    /// Exported under the `metadata` feature only.
+    #[cfg(feature = "metadata")]
+    #[inline(always)]
+    #[must_use]
+    pub fn iter_comments(&self) -> impl Iterator<Item = &str> {
+        self.comments.iter().map(|s| s.as_str())
+    }
+    /// _(metadata)_ Doc-comments defined on the custom type.
+    /// Exported under the `metadata` feature only.
+    #[cfg(feature = "metadata")]
+    #[inline(always)]
+    #[must_use]
+    pub fn comments(&self) -> &[crate::SmartString] {
+        &self.comments
+    }
 }
 
 /// _(internals)_ A collection of custom types.
@@ -42,14 +74,14 @@ impl CustomTypesCollection {
     /// Register a custom type.
     #[inline(always)]
     pub fn add(&mut self, type_name: impl Into<Identifier>, name: impl Into<Identifier>) {
-        self.add_raw(
-            type_name,
-            CustomTypeInfo {
-                display_name: name.into(),
-                #[cfg(feature = "metadata")]
-                comments: <_>::default(),
-            },
-        );
+        let type_name = type_name.into();
+        let custom_type = CustomTypeInfo {
+            type_id: type_name.clone(),
+            display_name: name.into(),
+            #[cfg(feature = "metadata")]
+            comments: <_>::default(),
+        };
+        self.add_raw(type_name, custom_type);
     }
     /// Register a custom type with doc-comments.
     /// Exported under the `metadata` feature only.
@@ -61,13 +93,13 @@ impl CustomTypesCollection {
         name: impl Into<Identifier>,
         comments: impl IntoIterator<Item = C>,
     ) {
-        self.add_raw(
-            type_name,
-            CustomTypeInfo {
-                display_name: name.into(),
-                comments: comments.into_iter().map(Into::into).collect(),
-            },
-        );
+        let type_name = type_name.into();
+        let custom_type = CustomTypeInfo {
+            type_id: type_name.clone(),
+            display_name: name.into(),
+            comments: comments.into_iter().map(Into::into).collect(),
+        };
+        self.add_raw(type_name, custom_type);
     }
     /// Register a custom type.
     #[inline(always)]
@@ -75,6 +107,7 @@ impl CustomTypesCollection {
         self.add_raw(
             type_name::<T>(),
             CustomTypeInfo {
+                type_id: type_name::<T>().into(),
                 display_name: name.into(),
                 #[cfg(feature = "metadata")]
                 comments: <_>::default(),
@@ -89,6 +122,7 @@ impl CustomTypesCollection {
         self.add_raw(
             type_name::<T>(),
             CustomTypeInfo {
+                type_id: type_name::<T>().into(),
                 display_name: name.into(),
                 #[cfg(feature = "metadata")]
                 comments: comments.iter().map(|&s| s.into()).collect(),